refactor: change long to double for amount
Refs: #196

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/DebitDto.java
@@ -10,6 +10,6 @@ import io.soabase.recordbuilder.core.RecordBuilder;
 public record DebitDto(
     UserListDto user,
     SplitBy splitBy,
-    Long value
+    Double value
 ) {
 }
